refactor: refactor strategies to spearate package
