+++ strategies/utils.go
@@ -1,0 +1,32 @@
+// Copyright 2025 Naren Yellavula
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package strategies
+
+// RemoveOverstrike removes the common overstrike pattern from strings
+func RemoveOverstrike(input string) string {
+	runes := []rune(input)
+	var output []rune
+
+	for i := 0; i < len(runes); i++ {
+		// Check if the current rune is part of an overstrike sequence
+		if i+2 < len(runes) && runes[i+1] == '\b' {
+			output = append(output, runes[i+2])
+			i += 2
+		} else {
+			output = append(output, runes[i])
+		}
+	}
+	return string(output)
+}
